+++ src/sciform/__init__.py
@@ -1,5 +1,3 @@
-__version__ = "0.27.2"
-
 from sciform.format_options import FormatOptions
 from sciform.formatter import Formatter
 from sciform.scinum import SciNum, SciNumUnc
@@ -13,7 +11,7 @@ from sciform.format_options import (
 from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
                            ExpMode, AutoExpVal, AutoDigits)
 
-__all__ = ['__version__', 'FormatOptions', 'Formatter', 'SciNum', 'SciNumUnc',
+__all__ = ['FormatOptions', 'Formatter', 'SciNum', 'SciNumUnc',
            'set_global_defaults', 'reset_global_defaults',
            'global_add_c_prefix', 'global_add_small_si_prefixes',
            'global_reset_si_prefixes', 'global_add_ppth_form',